added multiworld_row exchange algorithm from maddft_newexchange branch

+++ src/madness/chem/SCF.cc
@@ -1324,6 +1324,10 @@ vecfuncT SCF::apply_potential(World& world, const tensorT& occ,
 	  //if (world.rank() == 0) print("selecting exchange multi world");
 	  K.set_algorithm(Exchange<double,3>::Algorithm::multiworld_efficient);
 	}
+	else if (param.hfexalg()=="multiworld_row") {
+	  //if (world.rank() == 0) print("selecting exchange multi world row");
+	  K.set_algorithm(Exchange<double,3>::Algorithm::multiworld_efficient_row);
+	}
 	else if (param.hfexalg()=="largemem") {
 	  //if (world.rank() == 0) print("selecting exchange large memory");
 	  K.set_algorithm(Exchange<double,3>::Algorithm::large_memory);
@@ -1369,19 +1373,52 @@ vecfuncT SCF::apply_potential(World& world, const tensorT& occ,
     //     potentialmanager->apply_nonlocal_potential(world, amo, Vpsi);
     // }
 
+    // compute Vpsi and truncation
     START_TIMER(world);
+    const bool tile_Vpsi = true;
+    size_t min_tile = 10;
+    size_t ntile = std::min(amo.size(), min_tile);
     if (!molecule.parameters.pure_ae()) {
         gaxpy(world, 1.0, Vpsi, 1.0, gthpseudopotential->apply_potential(world, vloc, amo, occ, enl));
     } else {
-        gaxpy(world, 1.0, Vpsi, 1.0, mul_sparse(world, vloc, amo, vtol));
+        if (tile_Vpsi){
+            for (size_t ilo=0; ilo<amo.size(); ilo+=ntile) {
+                size_t iend = std::min(ilo+ntile,amo.size());
+                vecfuncT tmpamo(amo.begin()+ilo,amo.begin()+iend);
+                auto tmpVpsi = mul_sparse(world, vloc, tmpamo, vtol);
+
+                //truncate tmpVpsi
+                truncate(world, tmpVpsi);
+
+                //put the results into their final home
+                for (size_t i = ilo; i<iend; ++i){
+                    Vpsi[i] += tmpVpsi[i-ilo];
+                }
+            }
+            END_TIMER(world, "V*psi");
+        } else {
+            gaxpy(world, 1.0, Vpsi, 1.0, mul_sparse(world, vloc, amo, vtol));
+            END_TIMER(world, "V*psi");
+            START_TIMER(world);
+            truncate(world, Vpsi);
+            END_TIMER(world, "Truncate Vpsi");
+            print_meminfo(world.rank(), "Truncate Vpsi");
+        }
     }
 
-    END_TIMER(world, "V*psi");
+    //START_TIMER(world);
+    //if (!molecule.parameters.pure_ae()) {
+    //    gaxpy(world, 1.0, Vpsi, 1.0, gthpseudopotential->apply_potential(world, vloc, amo, occ, enl));
+    //} else {
+    //    gaxpy(world, 1.0, Vpsi, 1.0, mul_sparse(world, vloc, amo, vtol));
+    //}
 
-    START_TIMER(world);
-    truncate(world, Vpsi);
-    END_TIMER(world, "Truncate Vpsi");
-    print_meminfo(world.rank(), "Truncate Vpsi");
+    //END_TIMER(world, "V*psi");
+
+    //START_TIMER(world);
+    //truncate(world, Vpsi);
+    //END_TIMER(world, "Truncate Vpsi");
+    //print_meminfo(world.rank(), "Truncate Vpsi");
     world.gop.fence();
     return Vpsi;
 }
@@ -1520,23 +1557,67 @@ vecfuncT SCF::compute_residual(World& world, tensorT& occ, tensorT& fock,
     fpsi.clear();
     std::vector<double> fac(nmo, -2.0);
     scale(world, Vpsi, fac);
-    std::vector<poperatorT> ops = make_bsh_operators(world, eps);
-    set_thresh(world, Vpsi, FunctionDefaults<3>::get_thresh());
     END_TIMER(world, "Compute residual stuff");
 
-    START_TIMER(world);
-    vecfuncT new_psi = apply(world, ops, Vpsi);
-    END_TIMER(world, "Apply BSH");
-    ops.clear();
-    Vpsi.clear();
-    world.gop.fence();
+    const bool tile_applyBSH = true;
+    vecfuncT new_psi;
+
+    if (tile_applyBSH) {
+        START_TIMER(world);
+        size_t min_tile = 10;
+        size_t ntile = std::min(amo.size(), min_tile);
+        new_psi = zero_functions<double,3>(world, Vpsi.size());
+
+        for (size_t ilo=0; ilo<Vpsi.size(); ilo+=ntile) {
+            size_t iend = std::min(ilo+ntile,Vpsi.size());
+            vecfuncT tmp_Vpsi(Vpsi.begin()+ilo,Vpsi.begin()+iend);
+
+            int tmp_nmo = tmp_Vpsi.size();
+            tensorT tmp_eps(tmp_nmo);
+            for (int i = 0; i < tmp_nmo; ++i) {
+                tmp_eps(i) = std::min(-0.05, fock(i+ilo, i+ilo));
+            }
+
+            std::vector<poperatorT> ops = make_bsh_operators(world, tmp_eps);
+            set_thresh(world, tmp_Vpsi, FunctionDefaults<3>::get_thresh());
+
+            vecfuncT tmp_new_psi = apply(world, ops, tmp_Vpsi);
+
+            //truncate tmp_new_psi
+            truncate(world, tmp_new_psi);
+
+            //put the results into their final home
+            for (size_t i = ilo; i<iend; ++i){
+                new_psi[i] += tmp_new_psi[i-ilo];
+            }
+            ops.clear();
+        }
+
+        Vpsi.clear();
+        world.gop.fence();
+        END_TIMER(world, "Apply BSH");
+    } else {
+        START_TIMER(world);
+
+        std::vector<poperatorT> ops = make_bsh_operators(world, eps);
+        set_thresh(world, Vpsi, FunctionDefaults<3>::get_thresh());
+
+        new_psi = apply(world, ops, Vpsi);
+        
+        ops.clear();
+        Vpsi.clear();
+        world.gop.fence();
+
+        END_TIMER(world, "Apply BSH");
+        
+        START_TIMER(world);
+        truncate(world, new_psi);
+        END_TIMER(world, "Truncate new psi");
+    }
 
     // Thought it was a bad idea to truncate *before* computing the residual
     // but simple tests suggest otherwise ... no more iterations and
     // reduced iteration time from truncating.
-    START_TIMER(world);
-    truncate(world, new_psi);
-    END_TIMER(world, "Truncate new psi");
 
     START_TIMER(world);
     vecfuncT r = sub(world, psi, new_psi);
@@ -2062,28 +2143,69 @@ void SCF::solve(World& world) {
         //     //do_this_iter = false;
         //     param.maxsub = maxsub_save;
         // }
-
-        if (param.do_localize() && do_this_iter) {
-            START_TIMER(world);
-            Localizer localizer(world, aobasis, molecule, ao);
-            localizer.set_method(param.localize_method());
-            MolecularOrbitals<double, 3> mo(amo, aeps, {}, aocc, aset);
-            tensorT UT = localizer.compute_localization_matrix(world, mo, iter == 0);
-            UT.screen(trantol);
-            amo = transform(world, amo, transpose(UT));
-            truncate(world, amo);
-            normalize(world, amo);
-
-            if (!param.spin_restricted() && param.nbeta() != 0) {
-
-                MolecularOrbitals<double, 3> mo(bmo, beps, {}, bocc, bset);
+        const bool tile_localize = true;
+        if (tile_localize) {
+            if (param.do_localize() && do_this_iter) {
+                START_TIMER(world);
+                Localizer localizer(world, aobasis, molecule, ao);
+                localizer.set_method(param.localize_method());
+                MolecularOrbitals<double, 3> mo(amo, aeps, {}, aocc, aset);
                 tensorT UT = localizer.compute_localization_matrix(world, mo, iter == 0);
                 UT.screen(trantol);
-                bmo = transform(world, bmo, transpose(UT));
-                truncate(world, bmo);
-                normalize(world, bmo);
+
+                size_t min_tile = 10;
+                size_t ntile = std::min(amo.size(), min_tile);
+                vecfuncT new_amo = zero_functions<double,3>(world, amo.size());  
+
+                for (size_t ilo=0; ilo<amo.size(); ilo+=ntile){
+                    size_t iend = std::min(ilo+ntile,amo.size());
+                    auto U_slice = copy(transpose(UT)(_,Slice(ilo,iend-1)));
+
+                    auto tmp_amo = transform(world, amo, U_slice);
+
+                    truncate(world, tmp_amo);
+
+                    for (size_t i = ilo; i<iend; ++i){
+                        new_amo[i] += tmp_amo[i-ilo];
+                    }
+                }
+                normalize(world, new_amo);
+                amo = new_amo;
+
+                if (!param.spin_restricted() && param.nbeta() != 0) {
+
+                    MolecularOrbitals<double, 3> mo(bmo, beps, {}, bocc, bset);
+                    tensorT UT = localizer.compute_localization_matrix(world, mo, iter == 0);
+                    UT.screen(trantol);
+                    bmo = transform(world, bmo, transpose(UT));
+                    truncate(world, bmo);
+                    normalize(world, bmo);
+                }
+                END_TIMER(world, "localize");
             }
-            END_TIMER(world, "localize");
+        } else {
+            if (param.do_localize() && do_this_iter) {
+                START_TIMER(world);
+                Localizer localizer(world, aobasis, molecule, ao);
+                localizer.set_method(param.localize_method());
+                MolecularOrbitals<double, 3> mo(amo, aeps, {}, aocc, aset);
+                tensorT UT = localizer.compute_localization_matrix(world, mo, iter == 0);
+                UT.screen(trantol);
+                amo = transform(world, amo, transpose(UT));
+                truncate(world, amo);
+                normalize(world, amo);
+
+                if (!param.spin_restricted() && param.nbeta() != 0) {
+
+                    MolecularOrbitals<double, 3> mo(bmo, beps, {}, bocc, bset);
+                    tensorT UT = localizer.compute_localization_matrix(world, mo, iter == 0);
+                    UT.screen(trantol);
+                    bmo = transform(world, bmo, transpose(UT));
+                    truncate(world, bmo);
+                    normalize(world, bmo);
+                }
+                END_TIMER(world, "localize");
+            }
         }
 
         START_TIMER(world);
